merger module moved to icon-manifest folder
