+++ src/icon-manifest/index.ts
@@ -1,0 +1,4 @@
+export * from './defaultSchema';
+export * from './IconGenerator';
+export * from './languages';
+export * from './manifestMerger';
