+++ src/icon-manifest/manifestMerger.ts
@@ -1,7 +1,7 @@
 import * as _ from 'lodash';
 import * as path from 'path';
-import { IconGenerator } from '../icon-manifest/iconGenerator';
-import { extensions } from '../icon-manifest/supportedExtensions';
+import { IconGenerator } from './iconGenerator';
+import { extensions } from './supportedExtensions';
 import {
   IExtensionCollection,
   IFileExtension,
