+++ test/iconGenerator.test.ts
@@ -4,8 +4,7 @@ import * as fs from 'fs';
 import * as path from 'path';
 import * as _ from 'lodash';
 import { expect } from 'chai';
-import { schema as defaultSchema, schema } from '../src/icon-manifest/defaultSchema';
-import { IconGenerator } from '../src/icon-manifest/iconGenerator';
+import { schema as defaultSchema, schema, IconGenerator } from '../src/icon-manifest';
 import { extensions as files } from '../src/icon-manifest/supportedExtensions';
 import { extensions as folders } from '../src/icon-manifest/supportedFolders';
 import { FileFormat } from '../src/models';
